rag-engine/requirements.txt
fastapi
uvicorn
llama-index
psycopg[binary]
python-dotenv

